Move utils for actions into their own module, and fixed locationParams error
