+++ src/lib/actions/category.actions.ts
@@ -4,6 +4,7 @@ import { CreateCategoryParams } from '@/types/parameters.types';
 import { handleError } from '../utils';
 import { connectToDatabase } from '../database';
 import { Category, ICategory } from '@/lib/database/models';
+// import { documentToJson } from '@/lib/actions/utils.actions';
 
 export const createCategory = async ({
   categoryName,
@@ -23,7 +24,11 @@ export const getAllCategories = async (): Promise<ICategory[] | undefined> => {
   try {
     await connectToDatabase();
 
-    return Category.find().lean();
+    const categories = await Category.find();
+
+    // TODO! Why doesn't this work?
+    // return documentToJson<ICategory[]>(categories);
+    return JSON.parse(JSON.stringify(categories));
   } catch (error) {
     handleError(error);
   }
